(Mostly) working map
The map sprite covers almost all of the screen and some of the collisions when the POIs are more logically distributed still work, though in general need refactoring.
If the objects are set to the size of their sprite then the movement becomes extremely limited, so it has been commented out for now.