Fixed bug where clicking space bar interacts with every POI at once
Fixed bug where clicking space bar interacts with every POI all at once. The problem was that the margin of how close you had to be with a POI to interact with it was 10 times the length of the screen (which meant obviously whenever space bar was pressed all the POIs were in range).